+++ src/main/java/uk/gov/legislation/api/responses/Level.java
@@ -1,4 +1,4 @@
-package uk.gov.legislation.endpoints.document.responses;
+package uk.gov.legislation.api.responses;
 
 import io.swagger.v3.oas.annotations.media.Schema;
