+++ src/main/java/uk/gov/legislation/converters/DocumentMetadataConverter.java
@@ -3,7 +3,6 @@ package uk.gov.legislation.converters;
 import uk.gov.legislation.api.responses.CommonMetadata;
 import uk.gov.legislation.api.responses.DocumentMetadata;
 import uk.gov.legislation.endpoints.document.service.EffectsConverter;
-import uk.gov.legislation.endpoints.document.service.ExtentConverter;
 import uk.gov.legislation.transform.simple.Metadata;
 import uk.gov.legislation.util.Cites;
 import uk.gov.legislation.util.Types;
