+++ src/main/java/uk/gov/legislation/transform/simple/Level.java
@@ -23,8 +23,8 @@ public class Level {
 
     /* conversion */
 
-    uk.gov.legislation.endpoints.document.responses.Level convert() {
-        uk.gov.legislation.endpoints.document.responses.Level other = new uk.gov.legislation.endpoints.document.responses.Level();
+    uk.gov.legislation.api.responses.Level convert() {
+        uk.gov.legislation.api.responses.Level other = new uk.gov.legislation.api.responses.Level();
         other.name = this.name;
         other.id = this.id;
         other.href = Links.shorten(uri);
